chore: split config section into files.
