+++ pkg/config/linters.go
@@ -1,0 +1,11 @@
+package config
+
+type Linters struct {
+	Enable     []string
+	Disable    []string
+	EnableAll  bool `mapstructure:"enable-all"`
+	DisableAll bool `mapstructure:"disable-all"`
+	Fast       bool
+
+	Presets []string
+}
