+++ pkg/config/linters_settings_gocritic.go
