+++ pkg/config/linters_settings_gocritic_test.go
